requirements.txt
# Production dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database - Prisma ORM
prisma==0.15.0

# Redis & Celery
redis==5.2.1
celery==5.4.0

# Logging
structlog==24.4.0
python-json-logger==3.2.1
slowapi==0.1.9

# HTTP Client
httpx==0.28.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
